GeoSpatial_tools/requirements.txt
# Core Streamlit and web components
streamlit>=1.28.0
streamlit-folium>=0.15.0

# Data manipulation and analysis
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0

# Geospatial libraries
geopandas>=0.13.0
shapely>=2.0.0
rasterio>=1.3.0
fiona>=1.9.0
pyproj>=3.5.0
xarray>=2023.6.0
rioxarray>=0.15.0

# Mapping and visualization
folium>=0.14.0
leafmap>=0.22.0

# Machine learning and image processing
scikit-learn>=1.3.0
scipy>=1.11.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Deep learning (optional - for advanced features)
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
tensorflow-hub>=0.14.0

# Object detection
ultralytics>=8.0.0

# Utility libraries
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Additional geospatial tools
contextily>=1.3.0
earthengine-api>=0.1.360

# For improved performance
numba>=0.57.0
dask>=2023.8.0
